feat: add game compatition.
